Refactor: remove BookSearchListFragment and its UserIntent
