--- app/src/main/res/layout/activity_book_search.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/activity_book_search_nav_host_container"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:background="#C0C0C0"
-    tools:viewBindingIgnore="true"
-    />
-
